% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\name{.create_dist_from_sparse}
\alias{.create_dist_from_sparse}
\title{Calculate cosine distance from the sparse matrix}
\usage{
.create_dist_from_sparse(create_large_sparse_output)
}
\arguments{
\item{create_large_sparse_output}{output from create_large_sparse()}
}
\value{
square matrix
}
\description{
Calculate cosine distance from the sparse matrix
}
