% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\name{.create_large_sparse}
\alias{.create_large_sparse}
\title{Create sparse matrix from two sets of peak lists}
\usage{
.create_large_sparse(
  query_peak_list,
  subject_peak_list,
  lowerMassCutoff,
  upperMassCutoff,
  chunksize
)
}
\arguments{
\item{query_peak_list}{list of numeric vectors representing query peaks}

\item{subject_peak_list}{list of numeric vectors representing subject peaks}

\item{lowerMassCutoff}{masses below this number will not be considered}

\item{upperMassCutoff}{masses above this number will not be considered}

\item{chunksize}{numeric, bin spectrum into X-sized bins}
}
\value{
list of length 3, containing length of queries, length of subjects, and the sparse matrix
}
\description{
Create sparse matrix from two sets of peak lists
}
