% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\name{.trim_and_melt_dist}
\alias{.trim_and_melt_dist}
\title{Melt distance matrix into pairwise data.table}
\usage{
.trim_and_melt_dist(big_dist, create_large_sparse_output, similarity_cutoff)
}
\arguments{
\item{big_dist}{output from create_dist_from_sparse()}

\item{create_large_sparse_output}{output from create_large_sparse()}

\item{similarity_cutoff}{numeric to remove similarities above threshold, NA to return everythingg (default)}
}
\value{
pairwise distances as data.table
}
\description{
Melt distance matrix into pairwise data.table
}
