% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaster.R
\name{peak_blaster}
\alias{peak_blaster}
\title{MALDI-TOF MS peak "BLAST"er}
\usage{
peak_blaster(
  query_peak_list,
  subject_peak_list,
  lowerMassCutoff = 3000,
  upperMassCutoff = 15000,
  chunksize = 10,
  similarity_cutoff = NA
)
}
\arguments{
\item{query_peak_list}{list of numeric vectors representing query peaks}

\item{subject_peak_list}{list of numeric vectors representing subject peaks}

\item{lowerMassCutoff}{masses below this number will not be considered}

\item{upperMassCutoff}{masses above this number will not be considered}

\item{chunksize}{numeric, bin spectrum into X-sized bins}

\item{similarity_cutoff}{numeric to remove similarities above threshold, NA to return everythingg (default)}
}
\value{
data.table of pairwise similarities between query and subject peak lists
}
\description{
MALDI-TOF MS peak "BLAST"er
}
